tinder_playwright/requirements.txt
playwright
Pillow
#deepface
opencv-python
numpy
requests
keyboard
python-dotenv